Revert "Added orders page to merchant app (#256)"
This reverts commit 1ad8723655b940fd0f3f2e54c763d38e14d0d438.
