+++ apps/equifood-merchant/example.env
@@ -1,1 +1,1 @@
-API_URL = http://localhost:3333/api
+API_URL = http://localhost:3333
