+++ apps/equifood-merchant/src/app/layouts/CoreLayout.tsx
@@ -8,14 +8,12 @@ import { StackScreenProps } from '@react-navigation/stack';
 import AccountScreen from '../screens/AccountScreen/AccountScreen';
 import HomeScreen from '../screens/OrdersScreen/OrdersScreen';
 import MenuScreen from '../screens/MenuScreen/MenuScreen';
-import OrderListScreen from '../screens/OrderListScreen/OrderListScreen';
 import type { RootNavigationParams } from './RootLayout';
 
 export type CoreNavigationParams = {
   orders: undefined;
   menu: undefined;
   account: undefined;
-  orders: undefined;
 };
 
 export type CoreNavigationProps<T extends keyof CoreNavigationParams> =
@@ -52,7 +50,6 @@ function CoreLayout() {
     >
       <Tab.Screen name="orders" component={HomeScreen}></Tab.Screen>
       <Tab.Screen name="menu" component={MenuScreen}></Tab.Screen>
-      <Tab.Screen name="orders" component={OrderListScreen}></Tab.Screen>
       <Tab.Screen name="account" component={AccountScreen}></Tab.Screen>
     </Tab.Navigator>
   );
