--- apps/equifood-merchant/src/app/layouts/CoreLayout/CoreLayout.spec.tsx
@@ -1,9 +1,0 @@
-import { render } from '../../../test-utils/render';
-import CoreLayout from './CoreLayout';
-
-describe('Core Layout tests', () => {
-  test('starts on home screen', async () => {
-    const { getByTestId } = await render(<CoreLayout></CoreLayout>);
-    expect(getByTestId('home-screen')).toBeVisible();
-  });
-});
