--- apps/equifood-merchant/src/app/layouts/CoreLayout/CoreLayout.tsx
@@ -1,94 +1,0 @@
-import {
-  NavigationContainer,
-  createNavigationContainerRef,
-} from '@react-navigation/native';
-import {
-  BottomTabBar,
-  BottomTabBarProps,
-  createBottomTabNavigator,
-} from '@react-navigation/bottom-tabs';
-import { createStackNavigator } from '@react-navigation/stack';
-import HomeScreen from '../../screens/HomeScreen/HomeScreen';
-import OrderListScreen from '../../screens/OrderListScreen/OrderListScreen';
-import AccountScreen from '../../screens/AccountScreen/AccountScreen';
-//import Cart from '../../screens/Cart/Cart';
-import { Box, View } from 'native-base';
-import { CoreStackParams, CoreTabParams } from './CoreNavigatorParams';
-import OrderScreen from '../../screens/OrderScreen/OrderScreen';
-//import CartButton from '../../components/buttons/CartButton/CartButton';
-import Ionicons from '@expo/vector-icons/Ionicons';
-
-const Tab = createBottomTabNavigator<CoreTabParams>();
-const Stack = createStackNavigator<CoreStackParams>();
-
-const TabBar = (props: BottomTabBarProps) => (
-  <Box>
-    <BottomTabBar {...props} />
-  </Box>
-);
-
-function CoreNavigation() {
-  return (
-    <Tab.Navigator
-      tabBar={TabBar}
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ focused, color, size }) => {
-          let iconName: React.ComponentProps<typeof Ionicons>['name'];
-
-          if (route.name === 'account') {
-            iconName = focused ? 'md-person' : 'md-person-outline';
-          } else if (route.name === 'home') {
-            iconName = focused ? 'md-restaurant' : 'ios-restaurant-outline';
-          } /*else if (route.name === 'map') {
-            iconName = focused ? 'md-map' : 'md-map-outline';
-          }*/ else if (route.name === 'orders') {
-            iconName = focused ? 'md-list' : 'md-list-outline';
-          } else {
-            return;
-          }
-
-          // You can return any component that you like here!
-          return <Ionicons name={iconName} size={size} color={color} />;
-        },
-        tabBarActiveTintColor: 'green',
-        tabBarInactiveTintColor: 'gray',
-      })}
-    >
-      <Tab.Screen
-        name="home"
-        component={HomeScreen}
-        options={{ title: 'Restaurants' }}
-      />
-      <Tab.Screen
-        name="orders"
-        component={OrderListScreen}
-        options={{ title: 'Order List' }}
-      />
-      <Tab.Screen
-        name="account"
-        component={AccountScreen}
-        options={{ title: 'Account' }}
-      />
-    </Tab.Navigator>
-  );
-}
-function CoreLayout() {
-  const navigation = createNavigationContainerRef<CoreStackParams>();
-
-  return (
-    <View flex={1} testID="core-layout">
-      <NavigationContainer ref={navigation}>
-        <Stack.Navigator>
-          <Stack.Screen
-            name="core"
-            component={CoreNavigation}
-            options={{ headerShown: false }}
-          ></Stack.Screen>
-          <Stack.Screen name="order" component={OrderScreen}></Stack.Screen>
-        </Stack.Navigator>
-      </NavigationContainer>
-    </View>
-  );
-}
-
-export default CoreLayout;
