+++ libs/ui-shared/src/components/organisms/OrderView/OrderView.tsx
@@ -1,136 +1,92 @@
-import { Text, Divider, HStack, VStack, Box } from 'native-base';
-import { Item, Merchant, Order } from '@equifood/api-interfaces';
+import { Text, HStack, VStack, Box } from 'native-base';
+import { Item, Merchant } from '@equifood/api-interfaces';
 import { CheckoutItemCard } from '../../molecules/cards/CheckoutItemCard/CheckoutItemCard';
-import { FontAwesome } from '@expo/vector-icons';
-import { MerchantMap } from '../MerchantMap/MerchantMap';
 
 interface OrderViewProps {
-  order: Order;
-  viewHeight: number;
+  items: Item[];
+  quantities: { [itemId: string]: number };
+  merchant: Merchant;
 }
 
-export function OrderView({ order, viewHeight }: OrderViewProps) {
-  // computation goes here
-  const subtotal = order.items.reduce(
-    (acc, item) => acc + item.item.price * item.quantity,
-    0
-  );
-  const originalSubtotal = order.items.reduce(
-    (acc, item) => acc + item.item.originalPrice * item.quantity,
-    0
-  );
-
+export function OrderView({ items, quantities, merchant }: OrderViewProps) {
+  let totalPrice = 0;
+  for (const itemId in quantities) {
+    const item = items.find((item) => item.id === itemId);
+    const qt = quantities[itemId];
+    if (item) {
+      totalPrice += qt * item.price;
+    }
+  }
   return (
-    <Box>
-      <Box p="4">
-        <Text fontSize="md" pb="2">
-          Order Info
+    <VStack>
+      <HStack>
+        <Text fontSize="20" padding={3} flex={1}>
+          Pick up from
         </Text>
-        <Box width="full" flexDirection="column">
-          {order.items.map((item) => (
-            <Box key={item.item.id}>
-              <Box flexDirection="row" width="full" backgroundColor="red">
-                <Text>{`${item.quantity}x ${item.item.name}`}</Text>
-                <Text ml="auto" strikeThrough={true} color="red.600">
-                  ${(item.item.originalPrice * item.quantity).toFixed(2)}
-                </Text>
-                <Text ml="2">
-                  ${(item.item.price * item.quantity).toFixed(2)}
-                </Text>
-              </Box>
-              {item.item.description ? (
-                <Box pl="5">
-                  <Text lineBreakMode="tail">{item.item.description}</Text>
-                </Box>
-              ) : null}
-            </Box>
-          ))}
-        </Box>
-      </Box>
-
-      <Divider></Divider>
-      <Box p="4">
-        <Box flexDirection="row">
-          <Text color="red.600">Your savings</Text>
-          <Text color="red.600" ml="auto">
-            - ${(originalSubtotal - subtotal).toFixed(2)}
-          </Text>
-        </Box>
-        <Box flexDirection="row">
-          <Text>Subtotal</Text>
-          <Text ml="auto">${subtotal.toFixed(2)}</Text>
-        </Box>
-        <Box flexDirection="row">
-          <Text>GST</Text>
-          <Text ml="auto">${(subtotal * 0.05).toFixed(2)}</Text>
-        </Box>
-        <Box flexDirection="row">
-          <Text>Total</Text>
-          <Text ml="auto">${(subtotal * 1.05).toFixed(2)}</Text>
-        </Box>
-      </Box>
-      <Divider></Divider>
-      <Box p="4">
-        <Box>
-          <Text fontSize="md" pb="2">
-            Pickup Instructions
-          </Text>
-          <Text>
-            Please pickup at {order.merchant.location.address} before{' '}
-            {order.deadline.toLocaleDateString(undefined, {
-              hour: 'numeric',
-              minute: '2-digit',
-              day: 'numeric',
-              month: 'short',
-              year: 'numeric',
-            })}
-          </Text>
-          <Text pb="4">Payment: In-person</Text>
-          {viewHeight ? (
-            <MerchantMap
-              darkMode={false}
-              merchants={[order.merchant]}
-              initialRegion={{
-                longitude: order.merchant.location.longitude,
-                latitude: order.merchant.location.latitude,
-                latitudeDelta: 2,
-                longitudeDelta: 2,
-              }}
-              style={{ width: '100%', height: viewHeight * 0.5 }}
-            ></MerchantMap>
-          ) : null}
-        </Box>
-      </Box>
-      <Divider></Divider>
-      <Box p="4">
-        <Text>{`Placed ${order.order_date.toLocaleDateString(undefined, {
-          month: 'short',
-          day: 'numeric',
-          year: 'numeric',
-          minute: '2-digit',
-          hour: 'numeric',
-          timeZoneName: 'short',
-        })}`}</Text>
-      </Box>
-      <HStack p="5" flexDirection="row" alignItems="center" space="3">
-        <FontAwesome
-          name="recycle"
-          color="green"
-          style={{
-            fontSize: 50,
-          }}
-        ></FontAwesome>
         <Text
-          flexShrink={1}
-          fontSize="sm"
-          lineBreakMode="head"
-          textBreakStrategy="highQuality"
+          testID="restaurant"
+          fontWeight="extrabold"
+          fontSize="20"
+          paddingTop={3}
+          backgroundColor="green"
+          flex={2}
         >
-          Congratulations, you saved ${(originalSubtotal - subtotal).toFixed(2)}{' '}
-          today. Thank you for your contribution to reducing food waste across
-          Canada.
+          {merchant.name}
         </Text>
       </HStack>
-    </Box>
+      <HStack>
+        <Text fontSize="20" padding={3} flex={1}>
+          Located at
+        </Text>
+        <Text
+          fontWeight="bold"
+          testID="address"
+          fontSize="20"
+          alignSelf="center"
+          flex={2}
+        >
+          {merchant?.location?.address}
+        </Text>
+      </HStack>
+
+      <HStack>
+        <Text
+          fontSize="20"
+          paddingTop={3}
+          paddingLeft={3}
+          paddingRight={3}
+          flex={1}
+        >
+          By
+        </Text>
+        <Text fontSize="20" marginTop="3" fontWeight={'bold'} flex={2}>
+          Now - 15 min
+        </Text>
+      </HStack>
+
+      <VStack paddingTop={5}>
+        {(items || []).map((item) => (
+          <CheckoutItemCard
+            key={item.id}
+            item={item}
+            quantity={quantities[item.id] ?? 0}
+          ></CheckoutItemCard>
+        ))}
+      </VStack>
+      <HStack alignSelf={'flex-end'}>
+        <Text textAlign={'right'} alignSelf="center" fontSize={20}>
+          Total:
+        </Text>
+        <Text
+          testID="totalPrice"
+          fontWeight="bold"
+          fontSize="25"
+          alignSelf="center"
+          padding="3"
+        >
+          $ {totalPrice.toFixed(2)}
+        </Text>
+      </HStack>
+    </VStack>
   );
 }
