+++ libs/ui-shared/src/components/organisms/index.ts
@@ -1,4 +1,3 @@
 export * from './MerchantMap/MerchantMap';
-export * from './OrderConfirmView/OrderConfirmView';
-export * from './LoginView/LoginView';
 export * from './OrderView/OrderView';
+export * from './LoginView/LoginView';
